requirements.txt
##############################################
# Torch (CUDA 12.1)
##############################################
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

##############################################
# Ultralytics YOLO dependencies
##############################################
numpy
opencv-python
ultralytics
boto3
psycopg2-binary
logging


##############################################
# FastAPI, Web, HTTP
##############################################
fastapi
uvicorn[standard]
websockets
httpx
python-multipart

